+++ src/actions/ansible.py
@@ -73,7 +73,8 @@ class AnsiblePlaybookAction:
             'ansible-playbook',
             '-i', self.inventory,
             self.playbook,
-            '-e', f'ansible_host={target_host}'
+            '-e', f'ansible_host={target_host}',
+            '-e', f'ansible_user={config.ssh_user}'
         ]
 
         # Add resolved site-config vars first (extra_vars can override)
@@ -262,7 +263,7 @@ class EnsurePVEAction:
             'playbooks/pve-install.yml',
             '-e', f'ansible_host={target_host}',
             '-e', f'pve_hostname={self.pve_hostname}',
-            '-e', 'ansible_user=root',
+            '-e', f'ansible_user={config.ssh_user}',
         ]
 
         rc, out, err = run_command(cmd, cwd=ansible_dir, timeout=self.timeout)